Update locallib.php
We made this minor change to get around a conflict with New Relic's PHP Agent/JS beacon that looks for the closing tag. Without this fix, the JS conflict causes a fatal error disabling the print functionality. I realize this isn't a problem with this plugin per se, but I thought it might be helpful to other that are using New Relic or some other software that injects scripts based on finding the tag.